+++ src/components/Home.jsx
@@ -1,0 +1,48 @@
+import { Card, Typography } from "antd";
+import React from "react";
+
+const { Text } = Typography;
+
+const styles = {
+  text: {
+    fontSize: "16px",
+  },
+  card: {
+    boxShadow: "0 0.5rem 1.2rem rgb(189 197 209 / 20%)",
+    border: "1px solid #e7eaf3",
+    borderRadius: "0.5rem",
+  },
+};
+
+export default function Home() {
+  return (
+    <div style={{ display: "flex", gap: "10px" }}>
+      <Card
+        style={styles.card}
+        title={
+          <>
+            <Text strong>Gamify daily stretches
+              with AI and blockchain</Text>
+          </>
+        }
+      >
+        Have fun and get fit!
+      </Card>
+      <div>
+        <Card style={styles.card}>
+          <div>
+            <iframe
+              width="560"
+              height="315"
+              src="https://www.youtube.com/embed/y-SmsMRFeEc"
+              title="YouTube video player"
+              frameBorder="0"
+              allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture"
+              allowFullScreen>
+            </iframe>
+          </div>
+        </Card>
+      </div>
+    </div>
+  );
+}
